config: support yaml
merge yaml.go, json.go to parser.go
split `equal` from util.go to testutil.go
readme: add badges, yaml config example
