+++ fixture/config.yml
@@ -1,16 +1,22 @@
 rules:
-  - host: "localhost"
-  - path: "/api/v1/:type"
-  - to:
-    - host: "localhost:3001"
-    - path: "/api/:type"
-    - headers:
-      - "X-HP-A": "hello"
-      - "X-HP-B": "world"
-
-  - host: "localhost:3001"
-  - path: "/api/*"
-  - to:
-    - type: "origin"
-  - headers:
-    - "X-HP": "true"
+- host: localhost
+  path: "/api/v1/:type"
+  to:
+    host: localhost:3001
+    path: "/api/:type"
+  headers:
+    X-HP-A: hello
+    X-HP-B: world
+- host: localhost:3001
+  path: "/api/*"
+  to:
+    type: origin
+  headers:
+    X-HP: 'true'
+- host: localhost:4000
+  path: "*"
+  to:
+    host: httpbin.org
+    path: "/get"
+  headers:
+    X-HP: 'true'
