--- json.go
@@ -1,11 +1,0 @@
-package main
-
-import "encoding/json"
-
-func parseJSON(path string) Config {
-	data := readfile(path)
-	var c Config
-	json.Unmarshal(data, &c)
-
-	return c
-}
