+++ util.go
@@ -3,8 +3,6 @@ package main
 import "path/filepath"
 import "io/ioutil"
 import "strconv"
-import "reflect"
-import "fmt"
 import "os"
 
 func toInt(i interface{}) int {
@@ -49,19 +47,4 @@ func readfile(path string) []byte {
 	}
 
 	return bytes
-}
-
-// utils for test
-func equal(args ...interface{}) {
-	if len(args)%2 != 0 {
-		panic(fmt.Errorf("not matched args"))
-	}
-
-	step := len(args) / 2
-	for i := 0; i < step; i++ {
-		if !reflect.DeepEqual(args[i], args[i+step]) {
-			panic(fmt.Errorf("expect %v to equal %v", args[i], args[i+step]))
-		}
-	}
-
 }
