--- yaml.go
@@ -1,5 +1,0 @@
-package main
-
-func parseYaml(path string) {
-
-}
