--- yaml_test.go
@@ -1,8 +1,0 @@
-package main
-
-import "testing"
-
-func TestParseYaml(t *testing.T) {
-	file := "./fixture/config.yml"
-	parseYaml(file)
-}
